+++ framework/Source/CPTLegend.h
@@ -2,7 +2,9 @@
 
 /// @file
 
+@class CPTFill;
 @class CPTLegend;
+@class CPTLineStyle;
 @class CPTPlot;
 @class CPTTextStyle;
 
@@ -35,6 +37,22 @@ extern NSString *const CPTLegendNeedsReloadEntriesForPlotNotification;
 
 /// @name Drawing
 /// @{
+
+/** @brief @optional This method gives the delegate a chance to provide a custom swatch fill for each legend entry.
+ *  @param legend The legend.
+ *  @param idx The zero-based index of the legend entry for the given plot.
+ *  @param plot The plot.
+ *  @return The fill for the legend swatch or @nil to use the default @link CPTLegend::swatchFill swatchFill @endlink .
+ **/
+-(CPTFill *)legend:(CPTLegend *)legend fillForSwatchAtIndex:(NSUInteger)idx forPlot:(CPTPlot *)plot;
+
+/** @brief @optional This method gives the delegate a chance to provide a custom swatch border line style for each legend entry.
+ *  @param legend The legend.
+ *  @param idx The zero-based index of the legend entry for the given plot.
+ *  @param plot The plot.
+ *  @return The line style for the legend swatch border or @nil to use the default @link CPTLegend::swatchBorderLineStyle swatchBorderLineStyle @endlink .
+ **/
+-(CPTLineStyle *)legend:(CPTLegend *)legend lineStyleForSwatchAtIndex:(NSUInteger)idx forPlot:(CPTPlot *)plot;
 
 /** @brief @optional This method gives the delegate a chance to draw custom swatches for each legend entry.
  *