+++ documentation/changelog.markdown
@@ -11,6 +11,7 @@ To be determined.
 - **New**: Added properties to control momentum scrolling for the x- and y-directions separately.
 - **New**: Added an identifier and user info dictionary to animation operations.
 - **New**: Animations can now start when the value of the animated property enters the animated range.
+- **New**: Added legend delegate methods to customize the swatch fill and border line style for each legend entry.
 - **Changed**: Switched to a new algorithm for computing curved scatter plots.
 - **Changed**: Miscellaneous bug fixes and cleanup.
 - **Removed**: Removed the elastic global range properties from `CPTXYPlotSpace`. Turning on momentum scrolling now automatically allows elastic ranges, too.